if using trip_id_key in Timeline, fallback to _id
When getting a timeline of cleaned_places and expected_trips, we use trip_id_key to identify the expected_trips by the cleaned_trip they came from.
But if there are cleaned_untracked entries in this mix, we don't need to do this on those. So, if the trip_id_key is not present, we will just use the entry's _id via getId().