fix first composite trip of batch not being created
Previously we used any() to check if there were are any entries, and then looped over the entries.
But using any() with an iterator consumes the first item, so it would not actually be included when looping. Thus, we would be missing the first of the batch.
Instead, let's just loop altogether and count what we get so that we can log accordingly.