Create and link the timeline properly, similar to the `CLEAN_AND_RESAMPLE` stage
Before this, the confirmed trip and place creation was somewhat ad-hoc:
- we first had only confirmed trips, no timeline, so all the links were broken
- we then added confirmed place, but didn't make sure that links were all set properly
- we completely ignored confirmed untracked
- we switched to using a timeline to get alternating places and trips, but then
  we didn't update the last ending place when a new trip came in

This rewrites the code to follow a similar pattern to the `CLEAN_AND_RESAMPLE`
stage, which we know works.

This is a fix for https://github.com/e-mission/e-mission-docs/issues/880#issuecomment-1500543312

We find the last cleaned place and create or update the correspoding confirmed place.
We then iterate over the cleaned trips, update the start place, create a new
end place and link all three of them better.

We also included confirmed_untracked so that the full timeline is created
properly, which involved supporting the confirmed_untracked in multiple code
locations

Also switched the entry creation from new Entry to make_entry so that the
metadata is filled in correctly, similar to 418cffccbb58e2ac990c1d5a677ea2156c06abd9

Testing done:
- Before this change, when running the pipeline twice, the last place of the
  first run is not updated (https://github.com/e-mission/e-mission-docs/issues/880#issuecomment-1500543312)
- After this change, when running the pipeline twice, the last place of the first run is updated
    (https://github.com/e-mission/e-mission-docs/issues/880#issuecomment-1502750747)