Update composite trips when confirmed objects are updated
Before the composite trip changes, we retrieved confirmed trips for display.
So if/when the confirmed trips were updated, we would see them on the next phone retrieval.
In other words, we displayed the terminal output of the pipeline.

However, with the composite trip code, the terminal output is the composite trip.
But we update the confirmed trips as part of the matching.
So if we get new user inputs or update the confirmed place exit timestamps,
they will not be reflected in the display. We are no longer displaying the
terminal output of the pipeline, we have another display layer on top.

So when we update the confirmed objects, we need to figure out which the
matching composite objects are and update them as well.

We generally don't include forward references in our data objects, only backward references.
We already have a reference from the composite trip to the associated end_confirmed_place
We also add a reference to the associated confirmed_trip

Then, when we update a confirmed object, we find the related composite object
and update it as well.

This fixes: https://github.com/e-mission/e-mission-docs/issues/880#issuecomment-1500798479
https://github.com/e-mission/e-mission-docs/issues/880#issuecomment-1500799602
https://github.com/e-mission/e-mission-docs/issues/880#issuecomment-1503756634
https://github.com/e-mission/e-mission-docs/issues/880#issuecomment-1504109750

Testing done:

- Loaded data, no additions
- Added place additions via UI
- Pushed changes
- Re-ran pipeline

Additions match

```
>>> pd.json_normalize(all_places)["data.additions"]
0                                                   []
1                                                   []
2    [{'_id': 64361902f5622167bcbdf190, 'user_id': ...
3    [{'_id': 64361902f5622167bcbdf1c2, 'user_id': ...
4    [{'_id': 64361902f5622167bcbdf208, 'user_id': ...
5    [{'_id': 64361902f5622167bcbdf246, 'user_id': ...
6                                                   []
7                                                   []
8                                                   []
9                                                   []
```

Also in the composite trips

```
>>> pd.json_normalize(all_ct)["data.end_confirmed_place.data.additions"]
0                                                   []
1    [{'_id': 64361902f5622167bcbdf190, 'user_id': ...
2    [{'_id': 64361902f5622167bcbdf1c2, 'user_id': ...
3    [{'_id': 64361902f5622167bcbdf208, 'user_id': ...
4    [{'_id': 64361902f5622167bcbdf246, 'user_id': ...
5                                                   []
6                                                   []
7                                                   []
8                                                   []
```

Added trip inputs and additions via UI
Pushed changes
Re-ran pipeline; both show up in the confirmed trips

```
>>> pd.json_normalize(all_ct)["data.additions"]
0                                                   []
1                                                   []
2                                                   []
3                                                   []
4    [{'_id': 64361c6bf5622167bcbe08cf, 'user_id': ...
5    [{'_id': 64361c6bf5622167bcbe0921, 'user_id': ...
6                                                   []
7                                                   []
8                                                   []

>>> for ct in all_ct:
...     print(ct["data"]["user_input"])
...
{}
{}
{}
{}
{'trip_user_input':...}
{'trip_user_input':...}
{'trip_user_input':...}
{}
{}
```