Ensure that object <-> input matching code works properly
This fixes
https://github.com/e-mission/e-mission-docs/issues/880#issuecomment-1509568224

For the fix, we pass in whether the time type should be forced or inferred.
It is currently forced since all the inputs (trip input, trip addition, place addition) use `data.start_ts`.
If we do support place inputs in the future, we need to decide whether to use start/end or enter/exit
and change this code appropriately

Testing done:
- Added an automated test for this scenario
- Stepped through the test and confirmed that it worked properly
https://github.com/e-mission/e-mission-docs/issues/880#issuecomment-1510445897
- Automated test passes