Fix new issue uncovered in existing incremental test
Because of https://github.com/e-mission/e-mission-docs/issues/880#issuecomment-1509875714
we weren't actually trying to match all entries from the last place properly.
After fixing it, while running the existing test, we actually executed
`_get_next_cleaned_timeline_entry`, which failed because `starting_trip` was not a field in `Cleanedplace`

The check should not have been `tl_entry.data.end_place is not None` and
`tl_entry.data.starting_trip is not None`, which generate key errors, but
rather, `"end_place" in tl_entry.data`.

If we didn't find a matching cleaned trip, we also needed to check if we found
some untracked time instead

With these changes, the test passes