Use EPOCH_MININUM when the start timestamp is None
Before this, we returned EPOCH_MAXIMUM for the timestamp in case the end_ts was
None, as would happen for the last place in the timeline.

However, the first place in the timeline will have `enter_ts` as None. We
handle this by setting it to EPOCH_MINIMUM (aka 0) in both locations where we
were using EPOCH_MAXIMUM.

+ new logging statement to help debug this issue; currently commented out

Testing done:
Previously failing test passes https://github.com/e-mission/e-mission-server/pull/895#issuecomment-1510688293

This fixes
https://github.com/e-mission/e-mission-server/pull/895#issuecomment-1510548225