Manually upgrade the cryptography and wheel packages
I tried updating the base image to the emission dependencies, and it installed
scipy and updated requests, but it did not actually update cryptography.

So I force update cryptography and wheel.
And remove the now unused packages so they are not found by the scan

Image built with this passes all the scan checks