:back: Copy over the `ending_trip` also for the first place in the timeline
During the lazy migration, the first place in the timeline could be the first
place in this incremental batch, with confirmed trips before it. But those
confirmed trips don't have confirmed places.

So if the timeline.first_place has a cleaned `ending_trip`, we should replace
it with the corresponding confirmed trip

This fixes
https://github.com/e-mission/e-mission-docs/issues/898#issuecomment-1524081748

Testing done:
- re-ran the pipeline on the test users, did not have errors
- re-ran the pipeline on the durham dataset, did not have errors