:back: Fix the assertions while implementing the backwards compat hack for creating confirmed places
When we run the pipeline on an ongoing system, we may have a mix of old style
and new style confirmed trips. This is because the trips created on previous
runs will be "old style", but if there was unprocessed data before the current
run, they would have been converted to "new style" trips, along with linked
confirmed places, in the pipeline step before this one.

So we cannot blindly assert that, if we see an "old style" trip, there are no
confirmed places. Instead, we must assert that the start and end places are
likely to be cleaned places.

Even this is tricky, because of the "likely to be" issue.

Since the trips are linked, the start place of an "old style" trip is the end
place of the previous "old style" trip. Since we convert the data trip by trip,
we would have created the confirmed place that corresponds to the start place
of a trip as the end place of the previous trip. So we cannot add a blanket
check for start places

Further, if we did process some data in the previous CREATE_CONFIRMED_OBJECTS
stage, we would have created confirmed places for the last part of the
timeline. The first of those confirmed places would be the end place of an "old
style" confirmed trip. So we cannot add a blanket check for the end place either.

So we add more complex assertions for old-style trips:
- If it has a start confirmed place, the place should have been created as the
  end place while processing a previous composite trip
- if we have an end confirmed place, the place should be the start place of a
  "new style" confirmed trip

This fixes:
https://github.com/e-mission/e-mission-docs/issues/898#issuecomment-1523815853

Testing done:
- re-ran the pipeline on the test users, did not have errors
- re-ran the pipeline on the durham dataset, did not have errors