:zap: Avoid cursor timeouts by reading the confirmed trip list into memory
Before this change, we followed the high-level structure:
- open DB cursor
- for entry in cursor
    - process(entry)

If there were a lot of entries/the processing was slow, the cursor could time out in the for loop
In practice, the cursor timeout time was ~ 10 minutes

We improve the robustness of this codebase at the cost of additional memory
usage by reading all entries into memory. We read all trips into memory while
building the trip model, and while generating the public dashboard so this does
not seem to be an excessive memory burden.

And we do not anticipate reading a large number of trips after this initial
creation until the pipeline is reset

This fixes:
https://github.com/e-mission/e-mission-docs/issues/898#issuecomment-1526857742

Testing done:
- Re-ran pipeline on durham data; no failures