:bug: `get_first_place` and `get_last_place` support timelines with a single place
We can sometimes run into the case where the timeline for cleaned -> composite
conversion consists of a single place.

With the previous implementation of `get_first_place` and `get_last_place`, we
would return `None` if there were no trips, this causing the conversion to fail.

We now fix the methods so that they return the single place in this corner case.
Note that this is only the case in which there is a single place and no trips.
If there is even one trip, we will have two places (start and end)

This fixes:
https://github.com/e-mission/e-mission-docs/issues/898#issuecomment-1526862427
https://github.com/e-mission/e-mission-docs/issues/898#issuecomment-1526891589