:back: Run validation check for the next trip only when it is expected to exist
In 624082e71543f201e581af506deda956acbd17ba, we added a check that

> if we have an end confirmed place, the place should be the start place of a
"new style" confirmed trip

This is generally true EXCEPT if the confirmed place is the end of the confirmed timeline.
In that case, there is no next trip.

So we only check for the next trip if the place has a "starting_trip" field and
we know that it is not the end of the chain.

This fixes
https://github.com/e-mission/e-mission-docs/issues/898#issuecomment-1528932432