:bug: add missing changes in `export.py`
In 956b48cc1155aba4f1f89a0c2b3c5c775c80ef8f, we pulled the changes required to change `bju -> esj`
For the `extract_timeline_for_day_range_and_user`, part of the functionality is
pulled out into this module. We resolved the conflict in the file by making the
changes in both the file and this module, but forgot to commit this module.

Fixing it now...