✨ Store a summary of the inferred and cleaned sections in the confirmed trip
This should make it easy to compute the metrics using "count every trip".

@allenmichael099 implemented this in the "Error bars" branch
https://github.com/e-mission/e-mission-eval-private-data/pull/33
and uses it in his calculations

This improves that computation by computing the distance, duration and count
summaries, and doing so from a single pandas dataframe. It also computes the
summaries while creating the confirmed trip, so that we don't need to recompute
manually after the fact, an operation which is very, very slow.

Testing done:
- new test cases pass
- existing TestUserInput on real data generates

```
        "inferred_section_summary": {
            "distance": {},
            "duration": {},
            "count": {}
        },
        "cleaned_section_summary": {
            "distance": {
                "ON_FOOT": 1047.1630675866315
            },
            "duration": {
                "ON_FOOT": 792.4609999656677
            },
            "count": {
                "ON_FOOT": 1
            }
        },
```

Note that the test subsequently fails with the following backtrace

```
======================================================================
ERROR: testTripUserInput (__main__.TestUserInput)
----------------------------------------------------------------------
Traceback (most recent call last):
  File "/Users/kshankar/e-mission/gis_branch_tests/emission/tests/analysisTests/userInputTests/TestUserInput.py", line 169, in testTripUserInput
    self.checkConfirmedTripsAndSections(dataFile, ld, preload=True,
  File "/Users/kshankar/e-mission/gis_branch_tests/emission/tests/analysisTests/userInputTests/TestUserInput.p
y", line 127, in checkConfirmedTripsAndSections
    self.compare_confirmed_objs_result(confirmed_trips, expected_confirmed_trips, manual_keys=["trip_user_input"] if trip_user_inputs else None)
  File "/Users/kshankar/e-mission/gis_branch_tests/emission/tests/analysisTests/userInputTests/TestUserInput.py", line 76, in compare_confirmed_objs_result
    self.assertEqual(len(rt.data['additions']), len(et.data['additions']))
KeyError: 'additions'

----------------------------------------------------------------------
```

This is because the `emission/tests/data/real_examples/shankari_2016-06-20.expected_confirmed_trips.manual_trip_user_input` does not have any additions.

But the code always adds additions

```
    confirmed_object_data["additions"] = \
        esdt.get_additions_for_timeline_entry_object(ts, tce)
```

And checking a production instance that does not use time-use, I still see

```
>>> edb.get_analysis_timeseries_db().count_documents({"metadata.key": "analysis/confirmed_trip", "data.additions": {"$exists": True}})
107
>>> edb.get_analysis_timeseries_db().count_documents({"metadata.key": "analysis/confirmed_trip", "data.additions": {"$exists": False}})
0
```

So it seems like we need to fix the ground truth.

However, this should then be failing on master as well, and it is not. Is this test not running by default?
Let's push this obviously broken test and see if it fails.
Otherwise, we need to fix the CI and then fix the ground truth