✅ Enable user input tests
The user input tests failed while run manually, but the automated tests were passing.
https://github.com/e-mission/e-mission-server/pull/917#issuecomment-1596131961
https://github.com/e-mission/e-mission-server/pull/917/commits/b912aad4205c3562741964a1a186d23bbce78617

On checking further, it looks like the user input tests were not run as part of the automated tests.
Let's turn them on now.
If this commit is correct, we expect that they will fail.
We will then be able to fix them by updating the ground truth