✨ Utility script to change deployment names
While testing against real data, we sometimes pull production data and test against it.
But now that the deployment config is automatically inferred from the opcode,
using the existing opcodes to login will try to connect to the real server
instead of our local development environment

This allows us to rename all the opcodes to `dev-emulator-*` and ensure that we
can test against our local dev server.