:zap: Skip the future pipeline stages if we have new data
Without this change, even if we have no new entries, we run through the rest of
the pipeline. Some of the queries for the rest of the pipeline can be fairly
slow - taking ~ 30 secs to complete. With ~ 100 people per program, even if we
had no new data, we will take 100 * 30 = 3000 secs = 50 minutes to process all
the data.

Instead, let's do an early return if we have no new data.
This allows us to hopefully cut that runtime down significantly.
This is a partial fix for
https://github.com/e-mission/e-mission-docs/issues/942#issuecomment-1674108246

Testing done:
- Added new test case which:
    - runs the pipeline on an empty dataset and checks that none of the states are updated
    - runs the pipeline on a non-empty dataset with no new data and checks that none of the states are updated
- Test passes
    ```
    ----------------------------------------------------------------------
    Ran 1 test in 2.361s

    OK
    ```