:wrench: add a command line option to decide whether to end early if there is no new data
In 7e5aacb6af1f0b7cce2a2e537c1e227ae23ce64f, we configured the pipeline to
return early if there is no new data. That is good for performance, but can
have some limitations if we make changes to the pipeline to address errors -
the modified stage will not run until there is new data.

Just to be on the safe side, add a command line option to control this feature,
and plumb it through the script ('intake_multiprocess' -> 'scheduler' ->
'intake') with a default of 'False' to retain the original behavior.

This is a partial fix of
https://github.com/e-mission/e-mission-docs/issues/942#issuecomment-1674108246

We were then able to change the test to run the pipeline with 'False' instead
of running the test pipeline. This added the 'OUTPUT_GEN' step to the set of valid states.
After adding that, the test passes

'''
----------------------------------------------------------------------
Ran 1 test in 3.825s

OK
'''