:zap: Stop running the trip model for people who have stopped contributing trips
We currently build the trip model for every single user, regardless of whether
they are actively contributing data or not. But if they have stopped
contributing data, their model is the same, we don't need to rebuild it.

So we skip the model building if their last trip's end is too old

This is a partial fix for:
https://github.com/e-mission/e-mission-docs/issues/942#issuecomment-1674136328

Testing done:
- Commented out the actual model
- Loaded the Denver CASR data
- Spot checked both decisions

```
2023-08-13 19:23:11,719:INFO:building model for user U1
2023-08-13 19:23:11,722:DEBUG:last trip was at 2022-10-22T15:51:49.993526+00:00, three weeks ago was 2023-07-24T02:23:11.722757+00:00, gap = 9 months ago, skipping model building...

2023-08-13 19:23:11,722:INFO:building model for user U2
2023-08-13 19:23:11,726:DEBUG:last trip was at 2023-08-01T01:46:10.019082+00:00, three weeks ago was 2023-07-24T02:23:11.726590+00:00, gap is in a week, building model...
```