:zap: Match confirmed and composite trips using an indexed value
When we match the user inputs to existing confirmed trips, we "fill in" the
inputs and update the confirmed trip. Hwoever, now that we retrieve the
composite trips, we need to propagate those changes downstream as well.

We used to find the matching composite trip for a confirmed trip by searching
for the entry whose `data.confirmed_trip` field was the id of the confirmed trip.
However, the `data.confirmed_trip` field is not indexed in the database.

So on larger data collection efforts, such as Denver CASR, composite trip to
confirmed trip mapping was taking 10 **minutes**. Since the confirmed and
composite trips will always have the same start and end timestamps we can find
the entry that has the same timestamp instead to find the match.

This drops the access time from 10 minutes to ~ 2 minutes
https://github.com/e-mission/e-mission-docs/issues/942#issuecomment-1677998639

Once we resolve the slow indexing on DocumentDB, this has the potential to drop
even further.

Side changes (also documented in
https://github.com/e-mission/e-mission-docs/issues/942#issuecomment-1678001429)
- we don't override the start and end timestamps and locations anymore (since
  we don't anticipate filling them in anyway)
- changed the unit tests, which were changing the `start_ts` to change other
  fields instead since we now lookup by `start_ts`

Testing done:
`TestCompositeTripCreation.py` passes with the changes