TestNominiatm Changes
Removed unused imports.

Added a line that temporarily sets the query URL in eco to that of the environment variable, just for the purpose of testing. This way, nominatim.json can stay the same, while the testing environment variable can still be used and modified easily.

Also resolved hardcoding of nominatim URL per Shankari's request.