✅ Fix the exception to expect from `add_section_summary`
In 73414b9919d21fb9ff9db110ed981f7d36aa57b6, we added logging to track how long
the section summaries take. The log prints out the trip id as the first line of
the function

```
+    logging.debug(f"get_section_summary({cleaned_trip['_id']}, {section_key}) called")
```

So the error when we pass in a fake trip is now that the `_id` doesn't exist,
not that the `user_id` doesn't exist.

Fixed by changing the expected exception

Testing done:

```
----------------------------------------------------------------------
Ran 1 test in 0.247s

OK
```