✅ Register the users before running the pipeline tests
In 0403ae0ecc8985911451f16fac085fcf0d6328de, we start precomputing the pipeline
stage and storing it in this profile. This means that we need the user to have
a profile to begin with.

In the real world, users will always have profiles since we register them
during `profile/create`

In this test, we set the `testEmail` so that we register the users as part of
the test as well. This implies that we also:
- need to delete the profile_db() as part of cleanup
- use the version of the pipeline in `epi` so that it will run the pre-population code

Once we do that, the tests pass. However, when they were failing, we realized
that we can't compare two tuples with `assertAlmostEqual` since the assertion
can't subtract tuples to get the difference. Instead, we separate out the start
and end and compare them independently with `assertAlmostEqual`

Testing done:
both tests pass

```
----------------------------------------------------------------------
Ran 2 tests in 33.884s

OK
```