Query URL woes
nominatim-docker-test: Updating geofabrik url variable, in hopes that it will be able to be passed into a call in TestNominatim.py. I have looked into it more, and I'm not sure if GH actions allows this.

nominatim.py: Realized that some of my assertion errors might be related to the nominatim url environment variable not getting set properly locally, so comparing local url to the one produced in github actions.

TestNominatim,py: Changing expected result for get_filtered_place. Difference was due to my error.