Get method test
Verified with last test that the .get method is the preferred way to set the variable. Trying a few ways to print the variables to see how it looks; I'm nervous because the last attempt printed "None". Added a print without any secrets to see if it's setting these variables at all.