Moving envs
Trying a few more things before I give up on GH actions secrets. Moving the environment variables to be after the run command, as that's what I've seen in many examples. If that doesn't work, I'm going to try to move it to the jobs section of the workflow file.