Fix for "url" KeyError observed in public-dash redesign testing
The bug was triggered by a .gitignore-d conf/ directory in public-dash which had a db.conf file.
This was being loaded when docker-compose.dev.yml was used to test the dev version of public-dash.

This was occurring since there was a nested dictionary in the db.conf.sample and db.conf files while I had initially only stored the nested keys (url, result_limit). Since it was still reading from the file, it stored the nested dictionary format with timeseries as the parent key followed by (url, result_limit) as children.

Hence, fixing it by adding the same nested dictionary structure in the emission/core/config.py and emission/core/get_database.py