Artifact + Matrix - 1
Added build and push to have latest image available in docker hub that is generated with updated YAML file.
Removed join repo from workflow dispatch.

Added artifact so that this can be used in case of push trigger event in admin-dash and public-dash repos with the help of the python script fetch_runID.py in those repos, to get the latest run_ID for a specific branch with the artifact.

Workflow dispatch would still use the input parameters which contains the latest timestamp tag.