Cleanup changes
Getting the server changes ready for merge. I modified the yml jobs to run on master again, since this was switched off for testing.  This might trigger a cascade of events, and I think we all the other PRs are ready so we can see if their jobs are triggered too.