add 'geocoded_address' to cleanedplace
We already query the OSM / Nominatim API for places in order to reverse-lookup the location. We use this to extract a short "display_name" (e.g. "Vine Street, Cincinnati"). However, this is all we have been using it for; the rest of the response was discarded

For looking up the mix of energy on the grid at a location (https://github.com/e-mission/e-mission-docs/issues/954), we'd like to have the ZIP code.
In addition, other fields in the 'address' section of the response may prove useful down the line