♻️  Move the config to a different file name that makes more sense
So that we can consolidate all the backwards compat code in one place.

if there is a config file and the environment variable is set, we need to
decide which one wins. We will pull out the code into a common class to ensure
DRY. Once the backwards compat has been removed, this can be merged into
the individual config files