✅ Fix the expected text while checking for tokens
In 7f1be920098f644d863536a5e810343f2634fc28, we changed the DB configuration
to be based on environment variables. This changed the output text when
importing the module and launching the script. This change fixes the test
token tests that compare the output text with a baseline to reflect the new
expected text.