✅ Remove environment variables that are likely to be different across runs
The tests failed because we now print all environment variables, and the
`hostname` changes between runs. So comparing the output with the known good
output always fails. We strip out variables that are likely to change over
time (hostname, conda version...) to make this test more robust.

1. We technically only need the hostname right now, but will add the conda
version as well so that we don't get spurious failures when the versions
change
2. this was not an issue earlier because we read the values from the config
file. We now read environment variables, but that brings in variables that we
did not set. So this is a new issue that we need to resolve by stripping them
out for the baseline comparison.

```
$ ./e-mission-py.bash emission/tests/storageTests/TestTokenQueries.py
----------------------------------------------------------------------
Ran 21 tests in 23.591s

OK
```