Logging CONDA_DEFAULT_ENV to investigate failed tests
Tests failed in CI/CD.
On testing locally it was able to fetch the `emissiontest` name.
But in the CI/CD, it is defaulting to use the `emission` environment name.

I did test locally using the steps in the manual install workflow file to use the activate_tests.sh and setup_tests.sh scripts.

Now testing locally by running the docker-compose.tests.yml which is used in the test-with-docker workflow.