Corrected 1st log message for local emission environment
The tests failed since despite using ENV vars, it wasn’t picking up value.
This was because ENV vars are being deleted in setUp().
Added CONDA_DEFAULT_ENV to the skipped env vars.

However, this approach of fetching ENV vars programmatically in the assert statement isn’t correct.
This is because the code in get_config() itself uses ENV vars to set the ret_val which is the exact value as being fetched in setUp(). Thus the values will always match.

———

Then I thought of adding another log message for the `emission` environment name.
While testing this I found that the 1st message isn’t being used in the normal situations when testing locally.
This message is observed when the conf/storage/db.conf file is used instead of the sample file.
But in normal operation, the server repo doesn’t have any conf files to begin with.

Additionally, we now set the DB_HOST as db and WEB_SERVER_HOST as 0.0.0.0 in the Dockerfile by default.

Hence the assert log message should also contain `Config file not  found…` message like the other two messages.
It should also have DB_HOST set to Db like the 3rd message.