Simplified tests with two log messages for four test environments
The PATH values vary based on a user's local system and could have additional values which could fail the test in a local environment run without docker containers.

Why do we even need the PATH, PYTHONPATH values in the assertion log messages.
This is because we are printing the config in get_database.py which contains environment variables or config file values.

Now, the question is why to print the entire set of values if we are only using DB_HOST, DB_RESULT_LIMIT?

I am now printing only these values in the get_database.py. This has two benefits:
- Allows to eliminate variable environment variables like PATH that might vary in local testing environments for every user.
- Since PATH is no longer printed, helps avoid issue of 'apple security' values.
- Helps reduce the number of log messages the 4 testing environments: local without docker, local with docker, CI/CD without docker, CI/CD with docker. The two docker environments have the same log message; similarly for the two non-docker environments. Thus we only need two log messages in the assertion tests.